backend/requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# HTTP Client
httpx==0.26.0

# Database - MySQL
pymysql==1.1.0
sqlalchemy==2.0.25
cryptography==41.0.7

# Document Processing
PyPDF2==3.0.1
python-docx==1.1.0
python-pptx==0.6.23
reportlab==4.0.9
Pillow==10.2.0

# Data Validation
pydantic==2.5.3
pydantic-settings==2.1.0

# Utilities
python-dotenv==1.0.0

# AI (Optional)
# openai==1.10.0

